# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Breathe ESG – data ingestion and analyst review platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ingest/sap:
    post:
      operationId: ingestSap
      tags: [ingestion]
      summary: Ingest SAP fuel/procurement flat file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SapIngestInput"
      responses:
        "201":
          description: Batch created with parsed records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionBatch"
        "400":
          description: Invalid file
  /ingest/utility:
    post:
      operationId: ingestUtility
      tags: [ingestion]
      summary: Ingest utility electricity CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UtilityIngestInput"
      responses:
        "201":
          description: Batch created with parsed records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionBatch"
        "400":
          description: Invalid file
  /ingest/travel:
    post:
      operationId: ingestTravel
      tags: [ingestion]
      summary: Ingest corporate travel CSV
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TravelIngestInput"
      responses:
        "201":
          description: Batch created with parsed records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionBatch"
        "400":
          description: Invalid file
...
components:
  schemas:
...
    IngestionBatch:
      type: object
      required: [id, clientId, sourceType, status, filename, rowCount, errorCount, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        sourceType:
          type: string
          enum: [sap, utility, travel]
        status:
          type: string
          enum: [processing, completed, failed]
        filename:
          type: string
        rowCount:
          type: integer
        errorCount:
          type: integer
        errorDetails:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
...
    SapIngestInput:
      type: object
      required: [clientId]
      properties:
        clientId:
          type: integer
    UtilityIngestInput:
      type: object
      required: [clientId]
      properties:
        clientId:
          type: integer
    TravelIngestInput:
      type: object
      required: [clientId]
      properties:
        clientId:
          type: integer